Update main_function.py
I added the buttons for moves during the players turn, and a quit button. So far I have "Draw," "Switch," and "Discard." Each one of the buttons calls their associated function which we can edit later. I also capped the frames to 60 because it was causing issues with the detection of the buttons.