Update main_function.py
The images are placeholders for now, but I added a visual hand, deck, discard, and dealer hand. I also added functionality to the switch function which also visually updates the hand and top of the discard deck. I'll work on spacing and card size once I get ahold of the Sabacc cards.